+++ app/src/main/java/com/jeroenmols/brickeffect/effect/Effect.java
@@ -1,4 +1,4 @@
-package com.jeroenmols.brickeffect;
+package com.jeroenmols.brickeffect.effect;
 
 import android.graphics.Bitmap;
